chore/#160 : 디렉토리 분리
