+++ feature/mypage/src/main/java/com/boostcamp/mapisode/mypage/intent/ProfileEditIntent.kt
@@ -1,0 +1,12 @@
+package com.boostcamp.mapisode.mypage.intent
+
+import com.boostcamp.mapisode.ui.base.UiIntent
+
+sealed interface ProfileEditIntent : UiIntent {
+	data object Init : ProfileEditIntent
+	data class NameChanged(val nickname: String) : ProfileEditIntent
+	data class ProfileChanged(val url: String) : ProfileEditIntent
+	data object PhotopickerClick : ProfileEditIntent
+	data object EditClick : ProfileEditIntent
+	data object BackClick : ProfileEditIntent
+}
