+++ feature/mypage/src/main/java/com/boostcamp/mapisode/mypage/sideeffect/MypageSideEffect.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.mapisode.mypage
+package com.boostcamp.mapisode.mypage.sideeffect
 
 import com.boostcamp.mapisode.ui.base.SideEffect
 
@@ -7,10 +7,4 @@ sealed interface MypageSideEffect : SideEffect {
 	data class ShowToast(val messageResId: Int) : MypageSideEffect
 	data object NavigateToLoginScreen : MypageSideEffect
 	data object NavigateToEditScreen : MypageSideEffect
-}
-
-sealed interface ProfileEditSideEffect : SideEffect {
-	data object Idle : ProfileEditSideEffect
-	data class ShowToast(val messageResId: Int) : ProfileEditSideEffect
-	data object NavigateToMypage : ProfileEditSideEffect
 }
