+++ feature/mypage/src/main/java/com/boostcamp/mapisode/mypage/state/MypageState.kt
@@ -1,0 +1,10 @@
+package com.boostcamp.mapisode.mypage.state
+
+import com.boostcamp.mapisode.ui.base.UiState
+
+data class MypageState(
+	val isLoading: Boolean = true,
+	val name: String = "",
+	val profileUrl: String = "",
+	val showWithdrawalDialog: Boolean = false,
+) : UiState
