+++ feature/mypage/src/main/java/com/boostcamp/mapisode/mypage/state/ProfileEditState.kt
@@ -1,13 +1,6 @@
-package com.boostcamp.mapisode.mypage
+package com.boostcamp.mapisode.mypage.state
 
 import com.boostcamp.mapisode.ui.base.UiState
-
-data class MypageState(
-	val isLoading: Boolean = true,
-	val name: String = "",
-	val profileUrl: String = "",
-	val showWithdrawalDialog: Boolean = false,
-) : UiState
 
 data class ProfileEditState(
 	val isLoading: Boolean = true,
